remove double definitions in builders
